Rework modularization
